pyproject.toml
[project]
name = "audible"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.4",
    "matplotlib>=3.10.9",
    "nbformat>=5.10.4",
    "openenv-core",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.7.2",
    "torch>=2.11.0",
    "transformers>=5.6.2",
    "trl>=1.2.0",
]

[tool.uv.sources]
openenv-core = { git = "https://github.com/meta-pytorch/OpenEnv.git" }
